Rename settings files and separate settings GUI
