+++ src/settings/__init__.py
@@ -1,1 +1,1 @@
-from .options import *
+from .settings import *
